Create makeDemandwareRequest util
