+++ web/app/integration-manager/_demandware-connector/categories/commands.js
@@ -1,6 +1,5 @@
-import {makeRequest} from 'progressive-web-sdk/dist/utils/fetch-utils'
 import {urlToPathKey} from '../../../utils/utils'
-import {initDemandWareSession} from '../app/commands'
+import {makeDemandwareRequest} from '../utils'
 import {receiveCategory} from '../../categories/responses'
 import {receiveProductListProductData} from '../../products/responses'
 import {parseProductListData} from '../parsers'
@@ -11,42 +10,39 @@ export const fetchProductListData = (url) => (dispatch) => {
     const categoryIDMatch = /\/([^/]+)$/.exec(url)
     const categoryID = categoryIDMatch ? categoryIDMatch[1] : ''
     const urlPathKey = urlToPathKey(url)
-    return initDemandWareSession()
-        .then((headers) => {
-            const requestOptions = {
-                method: 'GET',
-                headers
-            }
-            makeRequest(`${API_END_POINT_URL}/categories/${categoryID}`, requestOptions)
-                .then((response) => response.json())
-                .then((responseJSON) => {
-                    dispatch(receiveCategory({
-                        [urlPathKey]: {title: responseJSON.name}
-                    }))
-                    if (responseJSON.parent_category_id !== 'root') {
-                        makeRequest(`${API_END_POINT_URL}/categories/${responseJSON.parent_category_id}`, requestOptions)
-                            .then((response) => response.json())
-                            .then((responseJSON) => {
-                                dispatch(receiveCategory({
-                                    [urlPathKey]: {parentName: responseJSON.name, parentHref: `/s/${SITE_ID}/${responseJSON.id}`}
-                                }))
-                            })
-                    }
-                })
-                .then(() => {
-                    makeRequest(`${API_END_POINT_URL}/product_search?expand=images,prices&q=&refine_1=cgid=${categoryID}`, requestOptions)
-                        .then((response) => response.json())
-                        .then(({hits}) => {
-                            const productListData = parseProductListData(hits)
-                            const categoryData = {
-                                products: Object.keys(productListData)
-                            }
+    const requestOptions = {
+        method: 'GET'
+    }
 
-                            dispatch(receiveProductListProductData(productListData))
+    return makeDemandwareRequest(`${API_END_POINT_URL}/categories/${categoryID}`, requestOptions)
+            .then((response) => response.json())
+            .then((responseJSON) => {
+                dispatch(receiveCategory({
+                    [urlPathKey]: {title: responseJSON.name}
+                }))
+                if (responseJSON.parent_category_id !== 'root') {
+                    makeDemandwareRequest(`${API_END_POINT_URL}/categories/${responseJSON.parent_category_id}`, requestOptions)
+                        .then((response) => response.json())
+                        .then((responseJSON) => {
                             dispatch(receiveCategory({
-                                [urlPathKey]: categoryData
+                                [urlPathKey]: {parentName: responseJSON.name, parentHref: `/s/${SITE_ID}/${responseJSON.id}`}
                             }))
                         })
-                })
-        })
+                }
+            })
+            .then(() => {
+                makeDemandwareRequest(`${API_END_POINT_URL}/product_search?expand=images,prices&q=&refine_1=cgid=${categoryID}`, requestOptions)
+                    .then((response) => response.json())
+                    .then(({hits}) => {
+                        const productListData = parseProductListData(hits)
+                        const categoryData = {
+                            products: Object.keys(productListData)
+                        }
+
+                        dispatch(receiveProductListProductData(productListData))
+                        dispatch(receiveCategory({
+                            [urlPathKey]: categoryData
+                        }))
+                    })
+            })
 }
